+++ CHANGES.md
@@ -93,7 +93,7 @@ The CHANGES.md file starts at this version.
 
 3.2.1-SNAPSHOT - 28 May 2014
 -------------------
-
+- Better circuit constraint
 
 ????
 -------------------